consomme: remove mesh RPC interface
Move this interface up to the net_consomme layer--there is no reason to
have the concept of RPCs in the consomme crate, since there are no
async tasks.
