+++ vm/devices/net/net_consomme/consomme/Cargo.toml
@@ -9,7 +9,6 @@ rust-version.workspace = true
 [dependencies]
 inspect.workspace = true
 inspect_counters.workspace = true
-mesh.workspace = true
 pal_async.workspace = true
 
 futures.workspace = true
